Remove MATLAB mathematical_program_test and dead MATLAB code
Closes #7482.
