+++ matlab/BUILD.bazel
@@ -2,7 +2,4 @@
 
 load("//tools/lint:lint.bzl", "add_lint_tests")
 
-add_lint_tests(cpplint_extra_srcs = glob([
-    "**/*.cc",
-    "**/*.h",
-]))
+add_lint_tests(cpplint_extra_srcs = ["call_matlab_client.cc"])
