+++ matlab/CMakeLists.txt
@@ -16,7 +16,9 @@ include(cmake/flags.cmake)
 include(cmake/functions.cmake)
 include(cmake/packages.cmake)
 
-configure_file(addpath_drake.m.in addpath_drake.m @ONLY)
+configure_file(addpath_drake.m addpath_drake.m COPYONLY)
+
+install(FILES addpath_drake.m DESTINATION share/drake/matlab)
 
 if(UNIX AND NOT APPLE)
   # On Linux, force targets to use system Protobuf shared library instead of
@@ -35,19 +37,10 @@ if(UNIX AND NOT APPLE)
   set(CMAKE_CXX_FLAGS "-Wl,-rpath,${PROTOBUF_LIBRARY_DIR} ${CMAKE_CXX_FLAGS}")
 endif()
 
-drake_matlab_add_mex(mex_util mex_util.cc mex_util.h)
-target_link_libraries(mex_util drake::drake Eigen3::Eigen)
-
 drake_matlab_add_mex(call_matlab_client call_matlab_client.cc)
-target_link_libraries(call_matlab_client
-  drake::drake mex_util protobuf::libprotobuf
-)
+target_link_libraries(call_matlab_client drake::drake protobuf::libprotobuf)
 
-install(FILES addpath_drake_install.m DESTINATION share/drake/matlab
-  RENAME addpath_drake.m
-)
-
-install(TARGETS call_matlab_client mex_util DESTINATION share/drake/matlab)
+install(TARGETS call_matlab_client DESTINATION share/drake/matlab)
 
 if(BUILD_TESTING)
   add_subdirectory(test)
