--- matlab/CPPLINT.cfg
@@ -1,1 +1,0 @@
-filter=-build/include
