+++ matlab/addpath_drake.m
@@ -1,0 +1,3 @@
+function [] = addpath_drake()
+
+addpath(fileparts(mfilename('fullpath')));
