+++ matlab/cmake/packages.cmake
@@ -14,9 +14,6 @@ endif()
 find_package(drake CONFIG REQUIRED
   NO_CMAKE_SYSTEM_PATH NO_SYSTEM_ENVIRONMENT_PATH
 )
-find_package(Eigen3 3.3.3 CONFIG REQUIRED
-  NO_CMAKE_SYSTEM_PATH NO_SYSTEM_ENVIRONMENT_PATH
-)
 find_package(Matlab ${MINIMUM_MATLAB_VERSION} MODULE REQUIRED
   COMPONENTS MAIN_PROGRAM MEX_COMPILER MX_LIBRARY
 )
