+++ matlab/test/CMakeLists.txt
@@ -7,7 +7,3 @@ drake_matlab_add_unit_test(NAME matlab:call_matlab_client_test
 set_tests_properties(matlab:call_matlab_client_test PROPERTIES REQUIRED_FILES
   "${PROJECT_BINARY_DIR}/call_matlab_client.${Matlab_MEX_EXTENSION}"
 )
-
-drake_matlab_add_unit_test(NAME matlab:mathematical_program_test
-  COMMAND mathematical_program_test
-)
